requirements.txt
langchain
huggingface_hub
langchain-core
langchain-community

langchain-openai
streamlit
python-dotenv